Added Tests for Aggregation2D and Fixed Small Bug
I added the tests for 2-dimensional aggregation-based diagrams and also
fixed a small bug in FunctionJob. In FunctionJob, nothing should be
done if no data was obtained during the aggregation. This happens, e.g.,
if you want to compute the variance over time but just have a single
run. However, instead of checking whether the variable functionData is
null, it was checked whether the whole experiment set is null (which
cannot be)...